+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: kkaczoro <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/08 11:36:17 by kkaczoro          #+#    #+#             */
-/*   Updated: 2022/03/08 23:37:18 by kkaczoro         ###   ########.fr       */
+/*   Updated: 2022/03/10 18:11:39 by kkaczoro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	*ft_memset(void *s, int c, size_t n)
 	int	i;
 
 	i = 0;
-	while (i < n)
+	while (i < (int)n)
 	{
 		((char *)s)[i] = c;
 		i++;
